+++ src/ApplicationCore/Entities/SearchIndex/PQDisplay.cs
@@ -13,7 +13,7 @@ namespace InventoryManagementSystem.ApplicationCore.Entities.SearchIndex
         
         public virtual string Id { get; set; }
         public string PriceQuoteOrderNumber { get; set; }
-        public float NumberOfProduct { get; set; }
+        public int NumberOfProduct { get; set; }
         public string CreatedDate { get; set; }
         public string Deadline { get; set; }
         public string CreatedByName { get; set; }
